Re-added sortBy parameter to /factoids because it does make sense eg. for modifiedBy. sortBy needs more discussion anyhow.
Change schema for responses of /persons to PersonsResponse.

Change schema for responses of /sources to SourcesResponse.

Change schema for responses of /statements to StatementsResponse.

Change schema of responses to /factoids, /persons, /statements and /sources: the property of the array of found objects ist now 'data' (and so consistent) for all endpoints